+++ src/main/java/uk/org/ukfederation/mda/dom/AttrVisitor.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2013 University of Edinburgh.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.org.ukfederation.mda.dom;
+
+import net.shibboleth.metadata.dom.DomElementItem;
+
+import org.w3c.dom.Attr;
+
+/**
+ * Provides a variation of the Visitor pattern for performing operations on
+ * DOM attributes which are part of {@link DomElementItem} items.
+ */
+public interface AttrVisitor {
+
+    /**
+     * Called on each {@link Attr} visited as part of the processing
+     * of a {@link DomElementItem}.
+     * 
+     * @param visited the {@link Attr} being visited.
+     * @param item the {@link DomElementItem} which is the context for the visit.
+     */
+    public void visitAttr(Attr visited, DomElementItem item);
+    
+}